requirements.txt
google-generativeai
matplotlib
python-dotenv